join RHIFeatureLevel
