+++ cpp/unreal_plugins/SpUnrealTypes/Source/SpUnrealTypes/SpWorld.h
@@ -13,7 +13,7 @@
 #include "SpCore/Assert.h"
 #include "SpCore/Unreal.h"
 
-#include "SpUnrealTypes/SpRHIFeatureLevel.h"
+#include "SpUnrealTypes/SpRHIGlobals.h" // ESpRHIFeatureLevel
 
 #include "SpWorld.generated.h"
